requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.0
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
altair==4.2.2
    # via streamlit
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
backports-zoneinfo==0.2.1
    # via
    #   pytz-deprecation-shim
    #   tzlocal
blinker==1.6.2
    # via streamlit
cachetools==5.3.0
    # via streamlit
certifi==2023.7.22
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via streamlit
cmake==3.26.3
    # via triton
contourpy==1.0.7
    # via matplotlib
cycler==0.11.0
    # via matplotlib
datasets==2.11.0
    # via -r requirements.in
decorator==5.1.1
    # via validators
dill==0.3.6
    # via
    #   datasets
    #   multiprocess
entrypoints==0.4
    # via altair
filelock==3.12.0
    # via
    #   huggingface-hub
    #   torch
    #   triton
fonttools==4.43.0
    # via matplotlib
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2023.4.0
    # via
    #   datasets
    #   fsspec
    #   huggingface-hub
gitdb==4.0.10
    # via gitpython
gitpython==3.1.41
    # via streamlit
huggingface-hub==0.14.1
    # via datasets
idna==3.4
    # via
    #   requests
    #   yarl
importlib-metadata==6.6.0
    # via streamlit
importlib-resources==6.1.1
    # via
    #   jsonschema
    #   matplotlib
jinja2==3.1.3
    # via
    #   altair
    #   pydeck
    #   torch
jsonschema==4.17.3
    # via altair
kiwisolver==1.4.4
    # via matplotlib
lit==16.0.2
    # via triton
markdown-it-py==2.2.0
    # via rich
markupsafe==2.1.2
    # via jinja2
matplotlib==3.7.1
    # via
    #   seaborn
    #   ultralytics
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.14
    # via datasets
networkx==3.1
    # via torch
numpy==1.24.3
    # via
    #   altair
    #   contourpy
    #   datasets
    #   matplotlib
    #   opencv-python
    #   pandas
    #   pyarrow
    #   pydeck
    #   scipy
    #   seaborn
    #   streamlit
    #   torchvision
    #   ultralytics
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   nvidia-cusolver-cu11
    #   torch
nvidia-cuda-cupti-cu11==11.7.101
    # via torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
nvidia-cufft-cu11==10.9.0.58
    # via torch
nvidia-curand-cu11==10.2.10.91
    # via torch
nvidia-cusolver-cu11==11.4.0.1
    # via torch
nvidia-cusparse-cu11==11.7.4.91
    # via torch
nvidia-nccl-cu11==2.14.3
    # via torch
nvidia-nvtx-cu11==11.7.91
    # via torch
opencv-python==4.7.0.72
    # via ultralytics
packaging==23.1
    # via
    #   datasets
    #   huggingface-hub
    #   matplotlib
    #   streamlit
pandas==1.5.3
    # via
    #   altair
    #   datasets
    #   seaborn
    #   streamlit
    #   ultralytics
pillow==10.0.1
    # via
    #   matplotlib
    #   streamlit
    #   torchvision
    #   ultralytics
pkgutil-resolve-name==1.3.10
    # via jsonschema
protobuf==3.20.3
    # via streamlit
psutil==5.9.5
    # via ultralytics
pyarrow==11.0.0
    # via
    #   datasets
    #   streamlit
pydeck==0.8.0
    # via streamlit
pygments==2.15.1
    # via rich
pympler==1.0.1
    # via streamlit
pyparsing==3.0.9
    # via matplotlib
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
    #   streamlit
pytz==2023.3
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via
    #   datasets
    #   huggingface-hub
    #   ultralytics
requests==2.31.0
    # via
    #   datasets
    #   fsspec
    #   huggingface-hub
    #   responses
    #   streamlit
    #   torchvision
    #   ultralytics
responses==0.18.0
    # via datasets
rich==13.3.4
    # via streamlit
scipy==1.10.1
    # via ultralytics
seaborn==0.12.2
    # via ultralytics
sentry-sdk==1.21.0
    # via ultralytics
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
streamlit==1.21.0
    # via -r requirements.in
sympy==1.11.1
    # via torch
thop==0.1.1.post2209072238
    # via ultralytics
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
torch==2.0.0
    # via
    #   thop
    #   torchvision
    #   triton
    #   ultralytics
torchvision==0.15.1
    # via ultralytics
tornado==6.3.3
    # via streamlit
tqdm==4.65.0
    # via
    #   datasets
    #   huggingface-hub
    #   ultralytics
triton==2.0.0
    # via torch
typing-extensions==4.5.0
    # via
    #   huggingface-hub
    #   rich
    #   streamlit
    #   torch
tzdata==2023.3
    # via pytz-deprecation-shim
tzlocal==4.3
    # via streamlit
ultralytics==8.0.82
    # via -r requirements.in
urllib3==1.26.17
    # via
    #   requests
    #   responses
    #   sentry-sdk
validators==0.20.0
    # via streamlit
watchdog==3.0.0
    # via streamlit
wheel==0.40.0
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-cupti-cu11
    #   nvidia-cuda-runtime-cu11
    #   nvidia-curand-cu11
    #   nvidia-cusparse-cu11
    #   nvidia-nvtx-cu11
xxhash==3.2.0
    # via datasets
yarl==1.9.2
    # via aiohttp
zipp==3.15.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
